% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_verapdf.R
\name{configure_veraPDF}
\alias{configure_veraPDF}
\title{Configure the location of veraPDF installation}
\usage{
configure_veraPDF()
}
\value{
Invisibly returns \code{TRUE} if veraPDF was configured successfully, otherwise \code{FALSE}.
}
\description{
Opens an RStudio directory selection dialog and searches the selected directory and its
subdirectories for the veraPDF executable. This is helpful when veraPDF is installed, but
'eatArchive' fails to find it.
}
