% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_archive.R
\name{create_archive}
\alias{create_archive}
\title{Create a directory of all files to be archived}
\usage{
create_archive()
}
\value{
An Excel file, documenting all files that were copied, their original location, their new location, and whether they were converted to another file format.
}
\description{
Create a directory of all files to be archived
}
\examples{
\dontrun{
# Check the function
create_archive()
}

}
