% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_add_sheet_with_style.R
\name{.add_sheet_with_style}
\alias{.add_sheet_with_style}
\title{Format the excel sheet}
\usage{
.add_sheet_with_style(wb, sheet_name, df, autofill_values = NULL)
}
\arguments{
\item{autofill_values}{}
}
\description{
Format the excel sheet
}
\keyword{internal}
