% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_combine_excel_sheets.R
\name{.combine_excel_sheets}
\alias{.combine_excel_sheets}
\title{Combine tables from multiple excel sheets in directory report}
\usage{
.combine_excel_sheets(path_to_directory_report)
}
\arguments{
\item{path_to_directory_report}{Character. Path to excel file resulting from \code{write_directory_report()} containing column
"Archive", indicating whether and where to archive the file.}
}
\description{
Combine tables from multiple excel sheets in directory report
}
\keyword{internal}
