% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_docx_to_zip.R
\name{.convert_docx_to_zip}
\alias{.convert_docx_to_zip}
\title{Title}
\usage{
.convert_docx_to_zip(docx_path, save_to)
}
\arguments{
\item{docx_path}{}
}
\description{
Title
}
\keyword{internal}
