% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_eml_to_txt.R
\name{.convert_eml_to_txt}
\alias{.convert_eml_to_txt}
\title{Converts Thunderbird email objects (eml-files) to txt-files}
\usage{
.convert_eml_to_txt(eml_path, save_to = dirname(eml_path))
}
\description{
Converts Thunderbird email objects (eml-files) to txt-files
}
\keyword{internal}
