% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_xlsm_to_csv.R
\name{.convert_xlsm_to_csv}
\alias{.convert_xlsm_to_csv}
\title{Convert all sheets of an Excel macro file to CSV}
\usage{
.convert_xlsm_to_csv(xlsm_path, save_to, overwrite = FALSE)
}
\arguments{
\item{xlsm_path}{Path to the Excel macro file.}

\item{save_to}{Directory where the CSV files should be written.}

\item{overwrite}{Logical, whether to overwrite existing CSV files. Default = FALSE.}
}
\value{
Invisibly returns a character vector of paths to the written CSV files.
}
\description{
Convert all sheets of an Excel macro file to CSV
}
\keyword{internal}
