% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_xlsx_to_csv.R
\name{.convert_xlsx_to_csv}
\alias{.convert_xlsx_to_csv}
\title{Convert all sheets of an Excel file to CSV}
\usage{
.convert_xlsx_to_csv(xlsx_path, save_to, overwrite = FALSE)
}
\arguments{
\item{xlsx_path}{Path to the Excel file.}

\item{save_to}{Directory where the CSV files should be written.}

\item{overwrite}{Logical, whether to overwrite existing CSV files. Default = FALSE.}
}
\value{
Invisibly returns a character vector of paths to the written CSV files.
}
\description{
Convert all sheets of an Excel file to CSV
}
\keyword{internal}
