% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fix_umlaut.R
\name{.fix_umlaut}
\alias{.fix_umlaut}
\title{Umlaute in Dateinamen ersetzen (ä→ae, ö→oe, ü→ue, ß→ss; Großbuchstaben: Ae/Oe/Ue)}
\usage{
.fix_umlaut(x)
}
\description{
Umlaute in Dateinamen ersetzen (ä→ae, ö→oe, ü→ue, ß→ss; Großbuchstaben: Ae/Oe/Ue)
}
\keyword{internal}
