% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_file_info.R
\name{.get_file_info}
\alias{.get_file_info}
\title{Scan directory and write table of file names and info}
\usage{
.get_file_info(dir, recursive = FALSE)
}
\arguments{
\item{dir}{A directory.}

\item{recursive}{Logical.}
}
\description{
Scan directory and write table of file names and info
}
\keyword{internal}
