% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_longest_common_path.R
\name{.longest_common_path}
\alias{.longest_common_path}
\title{Find the longest common path in a vector of paths}
\usage{
.longest_common_path(vec_paths)
}
\arguments{
\item{vec_paths}{}
}
\description{
Find the longest common path in a vector of paths
}
\examples{
\dontrun{
paths <- c(
"c:/a/b/c/d",
"c:/a/b/e/f",
"c:/a/b/g/h",
"c:/a/i/j/d"
)
.longest_common_path(paths)
}
}
\keyword{internal}
