% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDirectory.R
\name{writeDirectory}
\alias{writeDirectory}
\title{Title}
\usage{
writeDirectory(
  base_dir,
  output_file,
  autofill_values = NULL,
  exclude_folders = c("_Archive")
)
}
\arguments{
\item{exclude_folders}{}
}
\description{
Title
}
