% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_archive_from_directory.R
\name{create_archive_from_directory}
\alias{create_archive_from_directory}
\title{Create an archival directory of all files in a directory}
\usage{
create_archive_from_directory()
}
\description{
Create an archival directory of all files in a directory
}
\examples{
\dontrun{
create_archive_from_directory()
}

}
