% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_archive_from_directory.R
\name{create_archive_from_directory}
\alias{create_archive_from_directory}
\title{Create an archival directory of all files in a directory}
\usage{
create_archive_from_directory(
  path_to_working_directory,
  path_to_archive_directory,
  exclude_folders = "_Archive",
  convert = TRUE,
  overwrite = FALSE,
  csv = "csv",
  pdf_flavor = "2b"
)
}
\arguments{
\item{path_to_working_directory}{Character. Path to working directory which is to be archived.}

\item{path_to_archive_directory}{Character. Path indicating where to create the archival directory.}

\item{exclude_folders}{Character vector. Names of subfolders to exclude from processing.}

\item{convert}{Logical, indicating whether to convert file formats or not. Current files for conversion are:
\itemize{
\item xlsx --> csv: each sheet of an excel file is converted to a csv file
\item xlsm --> csv: each sheet of an excel macro file is converted to a csv file
\item sav  --> csv: SPSS data files are converted to two csv files each (data, metadata)
\item eml  --> txt: Thunderbird email objects (eml-files) are converted to txt-files
\item docx --> txt: word files are converted to txt files
\item doc --> txt: word files are converted to txt files
}}

\item{overwrite}{Logical, indicating whether to overwrite existing files in archival directory.}

\item{csv}{Character specifying the csv type
\itemize{
\item "csv" uses "." for the decimal point and "," for the separator
\item "csv2" uses "," for the decimal point and ";" for the separator, the Excel convention for CSV files in some Western European locales.
}}
}
\value{
Folder and documentation of all files that have been copied.
}
\description{
Create an archival directory of all files in a directory
}
