% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_archive_from_report.R
\name{create_archive_from_report}
\alias{create_archive_from_report}
\title{Create an archival directory based on Excel file}
\usage{
create_archive_from_report(
  path_to_directory_report,
  path_to_archive_directory,
  convert = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{path_to_directory_report}{Character. Path to Excel file resulting from \code{write_directory_report()} containing column "AIP", indicating whether and where to archive the file.}

\item{path_to_archive_directory}{Character. Path to new archival directory.}

\item{convert}{Logical, indicating whether to convert file formats or not.}

\item{overwrite}{Logical, indicating whether to overwrite existing files or not.}
}
\value{
Folder of all files that have been copied as indicated in column "AIP" of the directory report, as well as a csv-file documenting all files that have been copied, including their original directory, new directory, and the information on whether the file had been converted or not.
}
\description{
Create an archival directory based on Excel file
}
