% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_add_sheet_with_style.R
\name{.add_sheet_with_style}
\alias{.add_sheet_with_style}
\title{Format the excel sheet}
\usage{
.add_sheet_with_style(wb, sheet_name, df, autocomplete_values = NULL)
}
\arguments{
\item{wb}{An excel workbook.}

\item{sheet_name}{Name of the excel sheet.}

\item{df}{A data.frame.}

\item{autocomplete_values}{A character vector of values which should be offered as autocomplete.}
}
\description{
Format the excel sheet
}
\keyword{internal}
