% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_docx_to_txt.R
\name{.convert_docx_to_txt}
\alias{.convert_docx_to_txt}
\title{Converts Word files (docx and doc) to txt-files}
\usage{
.convert_docx_to_txt(docx_path, save_to = dirname(docx_path))
}
\arguments{
\item{save_to}{}
}
\description{
Converts Word files (docx and doc) to txt-files
}
\keyword{internal}
