% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_pdf_to_pdfa.R
\name{.convert_pdf_to_pdfa}
\alias{.convert_pdf_to_pdfa}
\title{Convert a PDF file to PDF/A file with a particular flavor (e.g. "2b")}
\usage{
.convert_pdf_to_pdfa(pdf_path, save_to, pdf_flavor = "2b")
}
\arguments{
\item{pdf_flavor}{}
}
\description{
Convert a PDF file to PDF/A file with a particular flavor (e.g. "2b")
}
\keyword{internal}
