% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convert_sav_to_csv.R
\name{.convert_sav_to_csv}
\alias{.convert_sav_to_csv}
\title{Converts sav files (GADS objects) to csv-files}
\usage{
.convert_sav_to_csv(sav_path, save_to = dirname(sav_path), csv = csv)
}
\arguments{
\item{csv}{}
}
\description{
Converts sav files (GADS objects) to csv-files
}
\keyword{internal}
