% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Scan}
\usage{
write_metadata(
  base_dir,
  output_file,
  autofill_values = NULL,
  exclude_folders = c("_Archive")
)
}
\arguments{
\item{base_dir}{Path to directory that is to be scanned}

\item{output_file}{Path and filename of}

\item{exclude_folders}{}
}
\description{
Scan
}
